tmux-kit: add directional pane swap (Swap Left/Right/Up/Down)
Swap a pane with its adjacent pane in any direction via `swap-pane -s/-t -d` (shown only when a neighbor exists; never changes the active pane). Refactor panes.tsx into focused component/lib modules.
